src/requirements.txt
streamlit
openpyxl
sentence-transformers
rapidfuzz
pandas
numpy
faiss-cpu
pyinstaller
